Preprocessing

Documentation

lit :: String -> FProp

clause :: [[Char]] -> FProp

dimacs :: FilePath -> IO (Set FProp, [VarProp])

(dimacs f) is the pair (ps,vs) where ps is the set of formulas wich corresponds to the formula in DIMACS format writed in the file f and vs is the list of variables wich occurs in any propositional formula.

formulas :: FilePath -> IO (Set FProp, [VarProp])

(formulas f) is the pair (ps,vs) where ps is the set of formulas wich corresponds to the formula in Prover9 format writed in the file f and vs is the list of variables wich occurs in any propositional formula.